refactor(wsLink): Moves transformer logic to wsClient.request
